@@ -1,16 +1,24 @@
 {
   "name": "sanity-remix-template",
   "sideEffects": false,
+  "type": "module",
+
   "scripts": {
     "build": "remix build",
-    "dev": "remix dev",
-    "lint": "eslint . --fix",
-    "typecheck": "tsc --noEmit"
+    "dev": "remix dev --manual -c \"npm run start\"",
+    "start": "wrangler pages dev --compatibility-date=2023-06-21 ./public",
+    "typecheck": "tsc"
   },
   "dependencies": {
+    "@remix-run/cloudflare": "^1.19.3",
+    "@remix-run/cloudflare-pages": "^1.19.3",
+    "@remix-run/css-bundle": "^1.19.3",
+    "@remix-run/react": "^1.19.3",
+    "isbot": "^3.6.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "@portabletext/react": "^3.0.0",
     "@remix-run/node": "^1.19.3",
-    "@remix-run/react": "^1.19.3",
     "@remix-run/serve": "^1.19.3",
     "@resvg/resvg-js": "^2.4.1",
     "@sanity/asset-utils": "^1.3.0",
@@ -23,8 +31,6 @@
     "eslint-plugin-prettier": "^4.2.1",
     "groq": "^3.10.1",
     "lucide-react": "^0.216.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "remix-utils": "^6.3.0",
     "sanity": "^3.11.1",
     "sanity-plugin-iframe-pane": "^2.1.1",
@@ -33,21 +39,23 @@
     "zod": "^3.19.1"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.19.3",
-    "@remix-run/eslint-config": "^1.19.3",
     "@tailwindcss/typography": "^0.5.9",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
     "@types/styled-components": "^5.1.26",
     "autoprefixer": "^10.4.11",
-    "eslint": "^8.20.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "inquirer": "^9.1.4",
     "postcss": "^8.4.16",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.2.0",
     "tailwindcss": "^3.1.8",
-    "typescript": "^5.0.4"
+    "@cloudflare/workers-types": "^3.19.0",
+    "@remix-run/dev": "^1.19.3",
+    "@remix-run/eslint-config": "^1.19.3",
+    "@types/react": "^18.0.35",
+    "@types/react-dom": "^18.0.11",
+    "eslint": "^8.38.0",
+    "typescript": "^5.0.4",
+    "wrangler": "^3.1.1"
   },
   "engines": {
     "node": "16.x"
