@@ -1,2 +1,3 @@
 /// <reference types="@remix-run/dev" />
-/// <reference types="@remix-run/node" />
+/// <reference types="@remix-run/cloudflare" />
+/// <reference types="@cloudflare/workers-types" />
